separated HotButton component to a separate file
